files/programs/dockerhub_publish/program/requirements.txt
PyYaml
paho-mqtt
requests
python-dateutil
